integration rajaongkir api

+++ database/migrations/2023_06_18_044346_create_products_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->string('slug')->unique();
             $table->string('short_desc');
             $table->text('desc');
-            $table->decimal('weight', 10, 2);
-            $table->decimal('price', 15, 2);
+            $table->decimal('weight', 10);
+            $table->integer('price');
             $table->integer('qty')->default(10);
             $table->text('image');
             $table->timestamps();
